Aula 3.6 - Consumindo outra api e melhorando questões de environment.

+++ controllers/pagamentos.js
@@ -81,8 +81,8 @@ module.exports = (app) => {
 
     app.post('/pagamentos/pagamento', (req, res) => {
 
-        req.assert("forma_de_pagamento", "Forma de pagamento é obrigatório").notEmpty();
-        req.assert("valor", "Valor é obrigatório e deve ser um decimal").notEmpty().isFloat();
+        req.assert("pagamento.forma_de_pagamento", "Forma de pagamento é obrigatório").notEmpty();
+        req.assert("pagamento.valor", "Valor é obrigatório e deve ser um decimal").notEmpty().isFloat();
 
         var erros = req.validationErrors();
 
@@ -92,7 +92,7 @@ module.exports = (app) => {
             return;
         }
 
-        let pagamento = req.body;
+        let pagamento = req.body.pagamento;
 
         if (!Object.keys(pagamento).length) {
             throw new Error('Nenhum pagamento enviado.');
@@ -115,25 +115,39 @@ module.exports = (app) => {
             pagamento.id = resultado.recordset[0].identity;
 
             console.log('Pagamento criado com sucesso.');
-            res.location('/pagamentos/pagamento/' + pagamento.id);
+            res.location('/pagamentos/pagamento/' + pagamento.id)
 
-            var response = {
-                dados_do_pagamento: pagamento,
-                links: [
-                    {
-                        href: `${req.protocol}://${req.hostname}:${process.env.PORT}/pagamentos/pagamento/${pagamento.id}`,
-                        rel: 'confirmar',
-                        method: 'PUT'
-                    },
-                    {
-                        href: `${req.protocol}://${req.hostname}:${process.env.PORT}/pagamentos/pagamento/${pagamento.id}`,
-                        rel: 'cancelar',
-                        method: 'DELETE'
-                    }
-                ]
+            if (pagamento.forma_de_pagamento == 'cartao') {
+                var cartao = req.body.cartao;
+
+                console.log('cartão', cartao);
+
+                var clienteCartoes = new app.servicos.clienteCartoes();
+
+                clienteCartoes.autoriza(cartao, (error, request, response, retorno) => {
+                    console.log(retorno);
+                    res.status(201).json(retorno);
+                    return;
+                })
+            } else {
+                var response = {
+                    dados_do_pagamento: pagamento,
+                    links: [
+                        {
+                            href: `${req.protocol}://${req.hostname}:${process.env.PORT}/pagamentos/pagamento/${pagamento.id}`,
+                            rel: 'confirmar',
+                            method: 'PUT'
+                        },
+                        {
+                            href: `${req.protocol}://${req.hostname}:${process.env.PORT}/pagamentos/pagamento/${pagamento.id}`,
+                            rel: 'cancelar',
+                            method: 'DELETE'
+                        }
+                    ]
+                }
+    
+                res.status(201).json(response);
             }
-
-            res.status(201).json(response);
         });
 
     })
